separate config file
